ragaai-scheduler-full-lc/requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
pandas>=2.2.2
openpyxl>=3.1.2
langchain>=0.2.0
langchain-community>=0.2.0
langgraph>=0.2.32
